// backend/enums/types/component.type.ts
export const enum ComponentType {
    OBSTACLE = "OBSTACLE",
    BREAKABLE = "BREAKABLE",
    UPDATE = "UPDATE",
    SCORE = "SCORE",
    SPEED = "SPEED",
    DOOR = "DOOR",
    DAMAGE = "DAMAGE",
    GRID = "GRID",
    HEALTH_BAR = "HEALTH_BAR",
    ON_KILL = "ON_KILL",
    ON_HIT_DAMAGE = "ON_HIT_DAMAGE"
}